+++ backend/canisters/community/api/src/main.rs
@@ -65,7 +65,6 @@ fn main() {
     generate_ts_method!(community, report_message);
     generate_ts_method!(community, reset_invite_code);
     generate_ts_method!(community, send_message);
-    generate_ts_method!(community, send_message_with_transfer);
     generate_ts_method!(community, set_member_display_name);
     generate_ts_method!(community, set_video_call_presence);
     generate_ts_method!(community, tip_message);