+++ backend/canisters/community/api/src/updates/mod.rs
@@ -67,7 +67,6 @@ pub mod remove_reaction;
 pub mod report_message;
 pub mod reset_invite_code;
 pub mod send_message;
-pub mod send_message_with_transfer;
 pub mod set_member_display_name;
 pub mod set_video_call_presence;
 pub mod start_video_call_v2;