+++ backend/canisters/group/api/src/updates/mod.rs
@@ -54,7 +54,6 @@ pub mod remove_reaction;
 pub mod report_message;
 pub mod reset_invite_code;
 pub mod send_message_v2;
-pub mod send_message_with_transfer;
 pub mod set_video_call_presence;
 pub mod start_video_call_v2;
 pub mod tip_message;