+++ backend/canisters/group/impl/src/updates/mod.rs
@@ -47,7 +47,6 @@ pub mod remove_participant;
 pub mod remove_reaction;
 pub mod report_message;
 pub mod send_message;
-pub mod send_message_with_transfer;
 pub mod set_video_call_presence;
 pub mod start_video_call;
 pub mod tip_message;