+++ backend/integration_tests/src/client/community.rs
@@ -44,7 +44,6 @@ generate_msgpack_update_call!(remove_member);
 generate_msgpack_update_call!(remove_member_from_channel);
 generate_msgpack_update_call!(remove_reaction);
 generate_msgpack_update_call!(send_message);
-generate_msgpack_update_call!(send_message_with_transfer);
 generate_msgpack_update_call!(start_video_call_v2);
 generate_msgpack_update_call!(tip_message);
 generate_msgpack_update_call!(toggle_mute_notifications);